Modal is just presentation component. Logic moved to the button component.
