+++ src/scripts/modules/transformations/react/components/ActivityMatchingButton.jsx
@@ -2,7 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import createReactClass from 'create-react-class';
 import classnames from 'classnames';
-import { Label } from 'react-bootstrap';
+import { Map, fromJS } from 'immutable';
+import { Badge } from 'react-bootstrap';
+import { Loader } from '@keboola/indigo-ui';
+import StorageApi from '../../../components/StorageApi';
 import ActivityMatchingModal from '../modals/ActivityMatchingModal';
 
 export default createReactClass({
@@ -14,16 +17,24 @@ export default createReactClass({
 
   getInitialState() {
     return {
+      matches: Map(),
+      isLoading: false,
       showModal: false
     };
+  },
+
+  componentDidMount() {
+    this.loadDataAndRunSearch();
   },
 
   render() {
     return (
       <a onClick={this.openModal} className={classnames({ 'text-muted': this.props.disabled })}>
-        <i className="fa fa-fw fa-align-justify" /> Activity Matching{' '}
-        <Label bsStyle="info">BETA</Label>
+        {this.renderIcon()} Activity Matching{' '}
+        {!this.state.isLoading && <Badge>{this.state.matches.count()}</Badge>}
         <ActivityMatchingModal
+          matches={this.state.matches}
+          isLoading={this.state.isLoading}
           transformation={this.props.transformation}
           tables={this.props.tables}
           show={this.state.showModal}
@@ -31,6 +42,14 @@ export default createReactClass({
         />
       </a>
     );
+  },
+
+  renderIcon() {
+    if (this.state.isLoading) {
+      return <Loader className="fa-fw" />;
+    }
+
+    return <i className="fa fa-fw fa-align-justify" />;
   },
 
   openModal() {
@@ -43,5 +62,47 @@ export default createReactClass({
 
   closeModal() {
     this.setState({ showModal: false });
+  },
+
+  loadDataAndRunSearch() {
+    this.setState({ isLoading: true });
+    StorageApi.getActivityMatchingData()
+      .then((data) => this.findMatches(fromJS(data || [])))
+      .finally(() => {
+        this.setState({ isLoading: false });
+      });
+  },
+
+  findMatches(data) {
+    const sources = this.props.transformation
+      .get('input')
+      .map((mapping) => mapping.get('source'))
+      .toSet()
+      .toList();
+    const tables = data
+      .filter((row) => sources.includes(row.get('table')))
+      .map((row) => row.get('usedIn'));
+
+    if (sources.count() > tables.count()) {
+      return;
+    }
+
+    const matches = tables
+      .flatten(1)
+      .filter((row) => row.get('rowId') !== this.props.transformation.get('id'))
+      .groupBy((row) => row.get('rowId'))
+      .filter((configuration) => configuration.count() === sources.count())
+      .sortBy((configuration) => {
+        const lastRun = configuration.first().get('lastRunAt');
+        return -1 * new Date(lastRun).getTime();
+      })
+      .sortBy((configuration) => {
+        const status = configuration.first().get('lastRunStatus');
+        if (status === 'success') return -1;
+        if (status === 'error' || status === 'terminated') return 1;
+        return 0;
+      });
+
+    this.setState({ matches });
   }
 });
